% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_checks.R
\name{assert_current_scope}
\alias{assert_current_scope}
\title{Assert Current Scope}
\usage{
assert_current_scope(data_table, metadata)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}

\item{metadata}{result of \code{\link{get_tracking_metadata}}}
}
\value{
filtered data list after entities that are out of scope have been removed. Data
is in tidyxl::xlsx_cells format where each record corresponds to one cell in source data.
}
\description{
Scope of the data.
}
\seealso{
Other scope: 
\code{\link{empty_cols}()},
\code{\link{empty_rows}()},
\code{\link{exclude_fields}()},
\code{\link{header_rows}()},
\code{\link{inconsistent_column_names}()},
\code{\link{missing_fields}()},
\code{\link{question_marks}()}
}
\concept{scope}
