% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_prep.R
\name{combine_sheets}
\alias{combine_sheets}
\title{Combine Sheets}
\usage{
combine_sheets(data_table, time_metadata, dataset_type)
}
\arguments{
\item{data_table}{Data frame returned by `read_digitized_data`.}

\item{time_metadata}{List returned by \code{\link{make_time_metadata}}.}

\item{dataset_type}{Type of dataset (e.g., Mortality, Birth)}
}
\description{
Combine Sheets
}
