% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R, R/date_utilities.R
\name{convert_to_date}
\alias{convert_to_date}
\alias{convert_vec_to_date}
\title{Convert to date}
\usage{
convert_to_date(y, m, d, period)

convert_vec_to_date(y, m, d, period)
}
\arguments{
\item{y}{string containing 4-digit year}

\item{m}{string containing month numbers (1-12) or `NA`.}

\item{d}{string containing day numbers (1-31) or `NA`.}

\item{period}{a character string indicating whether the years are the
start or end of a period. This argument is only available when `m` and `d`
are `NA`. Valid inputs are "start"(January 1) or "end"(December 31),
default is NULL.}
}
\value{
Date object either a valid date or NA_Date_
}
\description{
Convert date string to date data type.
}
\section{Functions}{
\itemize{
\item \code{convert_vec_to_date()}: Vectorized version of date conversion

}}
