% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utilities.R
\name{create_date_fields}
\alias{create_date_fields}
\title{Create date fields}
\usage{
create_date_fields(data_table)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}
}
\value{
`data_table` containing additional fields named `period_start_date` and `period_end_date`
}
\description{
Creates two new fields of class `Date` named `period_start_date` and `period_end_date` in
input table using existing fields containing date range information.
}
