% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{date_range_issues}
\alias{date_range_issues}
\title{Date range issues}
\usage{
date_range_issues(data_table)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}
}
\value{
all records in `data_table` that correspond to gaps and overlaps in weekly date range data
}
\description{
Identify gaps and overlaps in weekly date ranges in Excel source data sheets.
}
\seealso{
Other data_quality_issues: 
\code{\link{invalid_dates}()},
\code{\link{repeated_field_names}()},
\code{\link{unclassified_field_names}()}
}
\concept{data_quality_issues}
