% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_checks.R
\name{empty_cols}
\alias{empty_cols}
\title{Empty cols}
\usage{
empty_cols(data_table)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}
}
\value{
all records in `data_table` that correspond to entire Excel sheet columns with no data
}
\description{
Identify entire columns in Excel sheets that contain no numeric or text data. These columns
are outside the boundaries of the data table contained in the sheet but they may contain Excel
formatting that causes the data reading functionality to perceive these columns as containing data.
}
\seealso{
Other scope: 
\code{\link{assert_current_scope}()},
\code{\link{empty_rows}()},
\code{\link{exclude_fields}()},
\code{\link{header_rows}()},
\code{\link{inconsistent_column_names}()},
\code{\link{missing_fields}()},
\code{\link{question_marks}()}
}
\concept{scope}
