% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_checks.R
\name{exclude_fields}
\alias{exclude_fields}
\title{Exclude fields}
\usage{
exclude_fields(data_table)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}
}
\value{
all records in `data_table` that correspond to excluded fields
}
\description{
Temporarily exclude some fields that require additional work/thought on how
they should be incorporated in the tidy data set.
These include:
- parish total fields - need to determine how to relate parish groupings with specific parishes
- relative fields - fields that contain relative counts
- possibly need to remove fields with age data with no range (ends in a number)
}
\seealso{
Other scope: 
\code{\link{assert_current_scope}()},
\code{\link{empty_cols}()},
\code{\link{empty_rows}()},
\code{\link{header_rows}()},
\code{\link{inconsistent_column_names}()},
\code{\link{missing_fields}()},
\code{\link{question_marks}()}
}
\concept{scope}
