% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{filter_out_data_quality}
\alias{filter_out_data_quality}
\title{Data quality filter}
\usage{
filter_out_data_quality(data_table, metadata, filter_data = TRUE, skip = TRUE)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}

\item{metadata}{result of \code{\link{get_tracking_metadata}}}

\item{filter_data}{boolean flag to filter out data quality issues. For the 'repeated_field_name' data quality issue, 
records are always filtered to create unique records despite this flag to ensure later pipeline processing steps can function.}

\item{skip}{completely skip data quality checking and just return the input
dataset}
}
\value{
Returns input `data_table` with possibly additional unique records corresponding to the 'repeated_field_name' data quality issue
when `filter_data=FALSE`. If `filter_data=TRUE` function will return `data_table` after data quality issue records have been removed.
}
\description{
Create data quality issue report from input table with the option to filter out data
quality issues.
}
