% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{fix_tidy_data_columns}
\alias{fix_tidy_data_columns}
\title{Fix Tidy Data Columns}
\usage{
fix_tidy_data_columns(data_set)
}
\arguments{
\item{data_set}{processed output from one of the
data processing functions (e.g. \code{\link{process_mortality}})}
}
\description{
Fix Tidy Data Columns
}
