% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_checks.R
\name{header_rows}
\alias{header_rows}
\title{Header rows}
\usage{
header_rows(data_table)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}
}
\value{
all records in `data_table` that correspond to header rows
}
\description{
Identify entire rows in Excel sheets that contain header comments in at least one
of the cells. These rows appear at the top of the Excel sheets and are assumed to
start with a hash tag character.
}
\seealso{
Other scope: 
\code{\link{assert_current_scope}()},
\code{\link{empty_cols}()},
\code{\link{empty_rows}()},
\code{\link{exclude_fields}()},
\code{\link{inconsistent_column_names}()},
\code{\link{missing_fields}()},
\code{\link{question_marks}()}
}
\concept{scope}
