% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_checks.R
\name{inconsistent_column_names}
\alias{inconsistent_column_names}
\title{Inconsistent time column names}
\usage{
inconsistent_column_names(data_table)
}
\arguments{
\item{data_table}{table containing mortality data from tidyxl::xlsx_cells output}
}
\value{
all records in `data_table` that correspond to inconsistent time column
names
}
\description{
Identify sheets that contain time data formatted differently than all other sheets.
(One Excel file contains weekly time data with no ".from" or ".to" suffix for "year",
"month", and "day" columns.)
}
\seealso{
Other scope: 
\code{\link{assert_current_scope}()},
\code{\link{empty_cols}()},
\code{\link{empty_rows}()},
\code{\link{exclude_fields}()},
\code{\link{header_rows}()},
\code{\link{missing_fields}()},
\code{\link{question_marks}()}
}
\concept{scope}
