% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{invalid_dates}
\alias{invalid_dates}
\title{Invalid dates}
\usage{
invalid_dates(data_table)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}
}
\value{
all records that correspond to invalid dates in `data_table`.
}
\description{
Identifies data records that correspond to invalid dates in Excel source data sheets. Invalid dates
refer to all dates that are `NA_Date_` after `create_date_fields()` was executed. For example,
non-existent dates (i.e.`2000-02-30`) are invalid dates.
}
\seealso{
Other data_quality_issues: 
\code{\link{date_range_issues}()},
\code{\link{repeated_field_names}()},
\code{\link{unclassified_field_names}()}
}
\concept{data_quality_issues}
