% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_prep.R
\name{lbom_pre_processing}
\alias{lbom_pre_processing}
\title{LBoM Preprocessing}
\usage{
lbom_pre_processing(data, metadata)
}
\arguments{
\item{data}{Dataset to be preprocessed in IIDDA pipelines.}

\item{metadata}{Metadata in form returned by iidda::get_dataset_metadata}
}
\description{
LBoM Preprocessing
}
