% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utilities.R
\name{make_date_string}
\alias{make_date_string}
\title{Concatenate Year, Month, and Day Fields into a yyyy-mm-dd String.}
\usage{
make_date_string(y, m, d, period = NULL)
}
\arguments{
\item{y}{vector containing years}

\item{m}{vector containing month numbers (1-12) or `NA`.}

\item{d}{vector containing day-of-month numbers (1-31) or `NA`.}

\item{period}{a character string indicating whether the years are the
start or end of a period. This argument is only available when `m` and `d`
are `NA`. Valid inputs are "start"(January 1) or "end"(December 31),
default is NULL.}
}
\description{
Concatenate Year, Month, and Day Fields into a yyyy-mm-dd String.
}
