% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utilities.R
\name{make_report_path}
\alias{make_report_path}
\title{Make Report Path}
\usage{
make_report_path(metadata)
}
\arguments{
\item{metadata}{result of \code{\link{get_tracking_metadata}}}
}
\description{
Make the path to be used for storing reports and supporting output,
by modifying the path to the tidy data to be found in the associated
metadata.
}
