% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_prep.R
\name{make_time_metadata}
\alias{make_time_metadata}
\title{Make Time Metadata}
\usage{
make_time_metadata(data_table)
}
\arguments{
\item{data_table}{Data frame returned by `read_digitized_data`.}
}
\description{
Make Time Metadata
}
