% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_checks.R
\name{missing_fields}
\alias{missing_fields}
\title{Missing fields}
\usage{
missing_fields(data_table)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}
}
\value{
all records in `data_table` that correspond to entire Excel sheet columns with no field name
}
\description{
Identify columns in Excel sheets that contain numeric data with no associated field name. Given
there is no field name the tidy data splitting functionality incorrectly assigns this data.
}
\seealso{
Other scope: 
\code{\link{assert_current_scope}()},
\code{\link{empty_cols}()},
\code{\link{empty_rows}()},
\code{\link{exclude_fields}()},
\code{\link{header_rows}()},
\code{\link{inconsistent_column_names}()},
\code{\link{question_marks}()}
}
\concept{scope}
