% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process}
\alias{process}
\alias{process_population}
\alias{process_all_cause}
\alias{process_births}
\alias{process_plague}
\alias{process_mortality}
\title{Process Data}
\usage{
process_population(data_table, metadata)

process_all_cause(data_table, metadata)

process_births(data_table, metadata)

process_plague(data_table, metadata)

process_mortality(data_table, metadata)
}
\arguments{
\item{data_table}{table containing data in tidyxl::xlsx_cells format}

\item{metadata}{result of \code{\link{get_tracking_metadata}}}
}
\value{
table containing counts in long format (by sex if applicable) and
compact time metadata format, age data extracted from cause field names
(if applicable)
}
\description{
Process all data to convert from tidyxl::xlsx_cells format to long format.
The output of this function contains data provenance information that
allow one to trace each record back to a particular cell in the
digitized Excel data.
}
\section{Functions}{
\itemize{
\item \code{process_population()}: Process population data

\item \code{process_all_cause()}: Process all-cause mortality data

\item \code{process_births()}: Process birth data

\item \code{process_plague()}: Process plague data

\item \code{process_mortality()}: Process mortality data

}}
