% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_sheet}
\alias{process_sheet}
\title{Process Excel Sheet}
\usage{
process_sheet(
  sheet,
  col_time_metadata,
  address_time_metadata,
  numeric_time_metadata,
  data_type = c("mortality", "all-cause-mortality", "births", "plague", "population")
)
}
\arguments{
\item{sheet}{table containing one Excel sheet in tidyxl::xlsx_cells format}

\item{col_time_metadata}{time metadata fields with "col_" prefix}

\item{address_time_metadata}{time metadata fields with "address_" prefix}

\item{numeric_time_metadata}{time metadata fields with "numeric_" prefix}

\item{data_type}{TODO: describe data type}
}
\value{
processed sheet as a tibble
}
\description{
Process excel sheet to convert from tidyxl::xlsx_cells format to tidy long format
}
