% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_checks.R
\name{question_marks}
\alias{question_marks}
\title{Question marks}
\usage{
question_marks(data_table)
}
\arguments{
\item{data_table}{table containing mortality data from tidyxl::xlsx_cells output}
}
\value{
all records in `data_table` that correspond to cells with question marks
}
\description{
Identify cells that contain question marks. It is not clear what type of missing data
is being represented by a question mark.
}
\seealso{
Other scope: 
\code{\link{assert_current_scope}()},
\code{\link{empty_cols}()},
\code{\link{empty_rows}()},
\code{\link{exclude_fields}()},
\code{\link{header_rows}()},
\code{\link{inconsistent_column_names}()},
\code{\link{missing_fields}()}
}
\concept{scope}
