% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_prep.R
\name{read_source_RDS}
\alias{read_source_RDS}
\title{Read Source RDS}
\usage{
read_source_RDS(
  source_data_folder,
  data_category = list("mortality", "acm", "birth", "plague", "population",
    "unknown_data_category")
)
}
\arguments{
\item{source_data_folder}{top level folder where source RDS files are saved}

\item{data_category}{data category}
}
\value{
tibble containing all data in specified data category
}
\description{
Read in source RDS data
}
