% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_prep.R
\name{split_source_excel}
\alias{split_source_excel}
\title{Split Source Excel}
\usage{
split_source_excel(
  source_data_folder,
  intermediate_data_folder,
  data_type = c("mortality", "all-cause-mortality", "births", "plague", "population")
)
}
\arguments{
\item{source_data_folder}{folder where source data is saved.}

\item{intermediate_data_folder}{top level folder for output, resulting RDS output file
will be saved in intermediate_data_folder > data_type location.}

\item{data_type}{data type to save as RDS file. Valid inputs are
mortality,
all-cause-mortality,
births,
plague,
population.
The default is \code{mortality}.}
}
\value{
the combined data frame that is saved in an RDS file
}
\description{
Read in source .xlsx data files and perform some data pre-processing.
Splits data into data categories and saves resulting table as RDS file.
}
