% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{filter_out_data_quality}
\alias{filter_out_data_quality}
\title{Data quality filter}
\usage{
filter_out_data_quality(
  data_table,
  metadata,
  filter_data = TRUE,
  append_repeated_field_names = FALSE
)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}

\item{metadata}{result of \code{\link{get_tracking_metadata}}}

\item{filter_data}{boolean flag to filter out data quality issues}
}
\value{
Returns input `data_table` when `filter_data=FALSE`. If there are any
repeated field names in Excel sheets these field names will be appended with "_(cell address)"
to ensure field names are unique within Excel sheets. If `filter_data=TRUE` function will
return `data_table` after data quality issue records have been removed when .
}
\description{
Create data quality issue report from input table with the option to filter out data
quality issues.
}
