% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{repeated_field_names}
\alias{repeated_field_names}
\title{Repeated field names}
\usage{
repeated_field_names(data_table)
}
\arguments{
\item{data_table}{table containing data from tidyxl::xlsx_cells output}
}
\value{
all records in `data_table` that correspond to repeated field names
}
\description{
Field names that appear multiple times in Excel source data sheets. In order to have one
data point per each time period (row in the sheet), the decision was made 
(https://github.com/davidearn/data_work/issues/192) to select the maximum numeric value
for each record for these fields. Additionally, for each time period if both numeric and
non-numeric values exist we prioritize numeric values, and in the presence of numeric ties
we select only one of the numeric values. All records identified by these decisions are 
updated to have a new column number (that does not contain data in the digitized files)
so they can easily be identified, and original columns are excluded in the final data set.
}
\seealso{
Other data_quality_issues: 
\code{\link{date_range_issues}()},
\code{\link{invalid_dates}()},
\code{\link{unclassified_field_names}()}
}
\concept{data_quality_issues}
